+++ src/main/java/org/greencodeinitiative/creedengo/python/PythonRuleRepository.java
@@ -26,11 +26,9 @@ import org.sonarsource.analyzer.commons.RuleMetadataLoader;
 import java.util.Arrays;
 import java.util.List;
 
+public record PythonRuleRepository(SonarRuntime sonarRuntime) implements RulesDefinition, PythonCustomRuleRepository {
 
-public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRepository {
-
-    @SuppressWarnings("rawtypes") // not possible to make a correction because super class is like that
-    static final List<Class> ANNOTATED_RULE_CLASSES = Arrays.asList(
+    static final List<Class<?>> ANNOTATED_RULE_CLASSES = Arrays.asList(
             AvoidGettersAndSetters.class,
             AvoidGlobalVariableInFunctionCheck.class,
             AvoidSQLRequestInLoop.class,
@@ -39,6 +37,7 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
             AvoidUnoptimizedVectorImagesCheck.class,
             AvoidFullSQLRequest.class,
             AvoidListComprehensionInIterations.class,
+            DetectBadLoggingFormatInterpolation.class,
             DetectUnoptimizedImageFormat.class,
             AvoidMultipleIfElseStatementCheck.class,
             LoopInvariantStatementCheck.class,
@@ -53,7 +52,10 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
             DisableGradientForModelEval.class,
             StringConcatenation.class,
             PreferAppendLeft.class,
-            AvoidCreatingTensorUsingNumpyOrNativePython.class
+            AvoidCreatingTensorUsingNumpyOrNativePython.class,
+            AvoidWildcardImportsCheck.class,
+            AvoidExceptionsForControlFlowCheck.class,
+            UsingSlotsOnDataClasses.class
     );
 
     public static final String LANGUAGE = "py";
@@ -61,17 +63,11 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
     public static final String RESOURCE_BASE_PATH = "org/green-code-initiative/rules/python";
     public static final String REPOSITORY_KEY = "creedengo-python";
 
-    private final SonarRuntime sonarRuntime;
-
-    public PythonRuleRepository(SonarRuntime sonarRuntime) {
-        this.sonarRuntime = sonarRuntime;
-    }
-
     @Override
     public void define(Context context) {
         NewRepository repository = context.createRepository(REPOSITORY_KEY, LANGUAGE).setName(NAME);
         RuleMetadataLoader ruleMetadataLoader = new RuleMetadataLoader(RESOURCE_BASE_PATH, sonarRuntime);
-        ruleMetadataLoader.addRulesByAnnotatedClass(repository, (List) checkClasses());
+        ruleMetadataLoader.addRulesByAnnotatedClass(repository, checkClasses());
         repository.done();
     }
 
@@ -80,9 +76,9 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
         return REPOSITORY_KEY;
     }
 
-    @SuppressWarnings("rawtypes") // not possible to make a correction because super class is like that
     @Override
-    public List<Class> checkClasses() {
+    public List<Class<?>> checkClasses() {
         return ANNOTATED_RULE_CLASSES;
     }
+
 }